feat: Added ErrorHandling + changed API-Key old one does not work anymore + improved processing speed (calculated attributes)

+++ src/main/java/at/fh/bif/swen/tourplanner/persistence/entity/TourLog.java
@@ -39,14 +39,14 @@ public class TourLog {
     private Duration totalTime;
 
     @Column(name = "rating")
-    private double rating;
+    private int rating;
 
     @ManyToOne(fetch = FetchType.LAZY,optional = false)
     @JoinColumn(name = "tour_id", nullable = false)
     @JsonIgnore
     private Tour tour;
 
-    public TourLog(LocalDate date, String comment, int difficulty, double totalDistance, Duration totalTimeLong, double rating){
+    public TourLog(LocalDate date, String comment, int difficulty, double totalDistance, Duration totalTimeLong, int rating){
         this.date = date;
         this.comment = comment;
         this.difficulty = difficulty;
